CIRC-1091 remove fee fine processing strategies
