--- src/main/java/org/folio/circulation/resources/renewal/RenewalFeeProcessingStrategy.java
@@ -1,12 +1,0 @@
-package org.folio.circulation.resources.renewal;
-
-import java.util.concurrent.CompletableFuture;
-
-import org.folio.circulation.resources.context.RenewalContext;
-import org.folio.circulation.support.Clients;
-import org.folio.circulation.support.results.Result;
-
-public interface RenewalFeeProcessingStrategy {
-  CompletableFuture<Result<RenewalContext>> processFeesFines(
-    RenewalContext renewalContext, Clients clients);
-}
